Debugging missing parentElement issue
Changes to show that the timeout function is called with out-of-date values. Check in the Console how the value of `showSnippet` is set to `false` when the Next button is pressed, but the call to generateExercise is still called because the closure it is in retains
a value of `true` for showSnippet.

Logic/functions.js
===
* Add value of showSnippet and render to call to generateExercise
* Log these values after generateExercise is called

TaskSection.jsx
===
* Use `let` for showSnippet, so that its value can be changed
* Log render index and showSnippet value on each function call
* Change value of showSnippet locally before calling setShowSnippet

handleAcceptClick
---
* Check if showSnippet is true in timeout, before calling generateExercise

handleNextClick
---
* Set showSnippet to false locally, and log it
* Call setCountDownVisible(false)

GameContext.jsx
===
* Replace React's setShowSnippet with a custom function
* Use setTHEShowSnippet for the React useState
* Log the value and the render index when setShowSnippet is called